currency and random unit test

+++ backend/util/random.go
@@ -19,12 +19,12 @@ func RandomInt(min, max int64) int64 {
 }
 
 // RandomString generates a random string of length n
-func RandomString(n int) string {
+func RandomString(n uint8) string {
 	var sb strings.Builder
 
 	k := len(alphabet)
 
-	for i := 0; i < n; i++ {
+	for i := 0; i < int(n); i++ {
 		c := alphabet[rand.Intn(k)]
 
 		sb.WriteByte(c)
